+++ docs/CHANGE_LOG.md
@@ -1,5 +1,13 @@
 # Syntax.js - Change Log:
 
+## Version 0.3.0:
+- Added Code/Pre element rendering support.
+- Added options attribute "data-syntax-options" support, which will override the default options per code element rendered.
+- Fixed a class name setting issue that caused a random space to appear.
+- Added a new custom trigger "onCopy", which is fired when the "Copy" button for a syntax element is pressed.
+
+<br>
+
 ## Version 0.2.0:
 - Added a border around the code container (uses root variables for easy changing).
 - Increased the padding slightly for the "Copy" button.